requirements.txt
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
requests==2.9.1
